webrtcbin: Allow NULL promises in a few places
It is very common for the caller to not want to be notified when
the tasks for set-remote-description or set-local-description are done,
so allow passing NULL for the promise to signify that you do not want to
be notified.

The same is not allowed for "create-answer" and "create-offer" since
that would be nonsensical.